store report api errors and show login button
