+++ apps/fishing-map/features/reports/Report.tsx
@@ -8,6 +8,8 @@ import { AsyncReducerStatus } from 'utils/async-slice'
 import { useLocationConnect } from 'routes/routes.hook'
 import { selectReportCategory } from 'features/app/app.selectors'
 import { selectActiveHeatmapDataviews } from 'features/dataviews/dataviews.selectors'
+import { WorkspaceLoginError } from 'features/workspace/WorkspaceError'
+import { selectLocationDatasetId, selectLocationAreaId } from 'routes/routes.selectors'
 import { useFetchReportArea, useFetchReportVessel } from './reports.hooks'
 import ReportSummary from './ReportSummary'
 import ReportTitle from './ReportTitle'
@@ -38,10 +40,22 @@ export default function Report() {
     },
   ]
   const filteredCategoryTabs = categoryTabs.filter((tab) => dataviewCategories.includes(tab.id))
-  const { status: reportStatus } = useFetchReportVessel()
+  const { status: reportStatus, error: statusError } = useFetchReportVessel()
   const { data: areaDetail } = useFetchReportArea()
+  const datasetId = useSelector(selectLocationDatasetId)
+  const areaId = useSelector(selectLocationAreaId)
 
-  if (reportStatus === AsyncReducerStatus.Error) return <p>There was a error</p>
+  if (reportStatus === AsyncReducerStatus.Error) {
+    if (statusError.status >= 400 && statusError.status < 500) {
+      return (
+        <WorkspaceLoginError
+          title={t('errors.privateReport', 'This is a private report')}
+          emailSubject={`Requesting access for ${datasetId}-${areaId} report`}
+        />
+      )
+    }
+    return <p>There was a unkown error</p>
+  }
   if (reportStatus !== AsyncReducerStatus.Finished) return <Spinner />
 
   const handleTabClick = (option) => {
